Did Robotics Stuff
AAunnovatedAuto.java: perfected auto.

RRHardware.java: fixed positions.

TOHardware.java: created TOHardware.java (short for TeleOp Hardware) so that to change positions required for an action or the action itself it only needs to be changed in one place. This can be used in all teleop programs so that the different programs can have different controls for the same standardized actions.